refactor(storage): extract facts, graph, ask stores from db.rs (gated, verbatim)
Pure MOVE, zero behavior change. Three gate-touching storage domains lift
out of db.rs into storage/{facts,graph,ask}_store.rs (inherent impl Db
split); db.rs -1914 lines (23,540 -> 21,626), storage/mod.rs +3.

- ask_store: insert/list_assistant_interactions_visible + list_assistant_
  threads_visible (both gated via meeting_is_visible) — BYTE-IDENTICAL.
- graph_store: the self-assembling + full-brain graph block (20 methods:
  build_graph/build_full_graph/graph_edges_visible/entity_*_visible/
  full_graph_links/…) + graph type aliases + map fns. Every
  visibility_clause + NOT-EXISTS predicate unchanged.
- facts_store: the bitemporal facts + user-facts cluster (list_facts_
  visible, apply_fact_ops, the supersession family, list/search_user_facts
  _visible, list_open_facts_visible, …) + row mappers. The interleaved
  memory/voiceprint/people methods + the seal-purge _tx helpers + the
  detail-coupled string readers were LEFT in db.rs.

Gate logic is byte-identical to trunk — the only non-cosmetic deltas are
4 fn->pub(crate) promotions (count_full_graph_nodes_visible,
entity_meeting_mentions_visible, full_graph_links,
reopen_import_superseded_facts_tx — db.rs tests / delete_meeting still
call them cross-file). visibility_clause stays in db.rs (the shared gate
helper), already pub(crate). Tests stay in db.rs (count 250 = trunk).
Verified: cargo test --lib fact/graph/entity/assistant/supersession +
sealed 103/0 + visible 31/0 (the gated tests prove the gate still fires),
clippy --lib -D warnings clean, MSRV grep clean.
