+++ src-tauri/src/storage/db.rs
@@ -10,19 +10,15 @@ use std::collections::HashSet;
 
 use crate::embed::Embedder;
 use crate::storage::models::{
-    Analytics, AssistantInteraction, AssistantThreadRow, BacklinkSource, Commitment,
+    Analytics, BacklinkSource, Commitment,
     CorrectionRecord, DayCount,
-    DocChunkHit, DocOutlineEntry, DocumentInfo, DocumentSummary, EntityDetail, EntityKind,
-    EntityNeighbor,
-    FullGraphData, FullGraphEdge, FullGraphEdgeKind, FullGraphNode, FullGraphNodeKind, FullGraphOpts,
-    GraphData,
-    GraphEdge, GraphEntity, GraphNode, Meeting, MeetingActionSummary, MeetingStatus, NoteCitation,
+    DocChunkHit, DocOutlineEntry, DocumentInfo, DocumentSummary, EntityKind,
+    GraphNode, Meeting, MeetingActionSummary, MeetingStatus, NoteCitation,
     NoteRecord, NoteSummary,
     OrgChunkHit, PendingShareAccept, PeopleList, PersonCard, PropertyKind,
     PropertyValue, RecipeRecord, SavedView, SearchHit,
     TypedNoteRow,
     StatusCount,
-    VaultSource,
 };
 use crate::transcribe::types::Segment;
 
@@ -98,7 +94,7 @@ impl FromStr for MeetingStatus {
 impl EntityKind {
     /// Stable lowercase string used as the on-disk `entities.kind` column value.
     /// Kept in sync with the serde `rename_all = "camelCase"` on the enum.
-    fn as_str(&self) -> &'static str {
+    pub(crate) fn as_str(&self) -> &'static str {
         match self {
             EntityKind::Person => "person",
             EntityKind::Project => "project",
@@ -8742,157 +8738,6 @@ impl Db {
         Ok(out)
     }
 
-    /// Persist one in-meeting voice-assistant interaction (the Q&A) against `meeting_id`. PURELY
-    /// ADDITIVE: it never touches note/transcript/timeline plaintext or blobs, so it can never blank
-    /// or clobber sealed content. `citations` is stored as a JSON array string. The CALLER is the
-    /// off-thread voice dispatch, which is best-effort + panic-free — a persist failure there is
-    /// logged (non-PII) and never disrupts recording/dispatch. Derived convenience data: it is PURGED
-    /// (not sealed) when the meeting's folder is sealed (see `purge_assistant_interactions_tx`).
-    #[allow(clippy::too_many_arguments)]
-    pub fn insert_assistant_interaction(
-        &self,
-        meeting_id: &str,
-        command: &str,
-        answer: &str,
-        citations: &[String],
-        status: &str,
-        source_label: Option<&str>,
-        thread_id: Option<&str>,
-        anchor_text: Option<&str>,
-        created_at: &str,
-    ) -> Result<i64> {
-        let citations_json = serde_json::to_string(citations)
-            .map_err(|e| AppError::Storage(format!("serialize interaction citations: {e}")))?;
-        let conn = self.lock();
-        conn.execute(
-            "INSERT INTO assistant_interactions
-                (meeting_id, command, answer, citations, status, source_label,
-                 thread_id, anchor_text, created_at)
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9)",
-            rusqlite::params![
-                meeting_id,
-                command,
-                answer,
-                citations_json,
-                status,
-                source_label,
-                thread_id,
-                anchor_text,
-                created_at
-            ],
-        )
-        .map_err(map_err)?;
-        Ok(conn.last_insert_rowid())
-    }
-
-    /// Read every persisted assistant interaction for `meeting_id`, oldest first, ONLY when the
-    /// meeting is VISIBLE to the session `unlocked` set (a sealed-and-not-session-unlocked meeting
-    /// returns an EMPTY list — never its rows). This is the gated read the detail DTO uses; it mirrors
-    /// the `meeting_is_visible` predicate so the interaction log is gated exactly like the note /
-    /// segments / timeline. Note: on seal the rows are PURGED (see `purge_assistant_interactions_tx`),
-    /// so a sealed meeting has no rows to read anyway — the gate is defense-in-depth.
-    pub fn list_assistant_interactions_visible(
-        &self,
-        meeting_id: &str,
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<AssistantInteraction>> {
-        if !self.meeting_is_visible(meeting_id, unlocked)? {
-            return Ok(Vec::new());
-        }
-        let conn = self.lock();
-        let mut stmt = conn
-            .prepare(
-                "SELECT command, answer, citations, status, source_label, created_at
-                   FROM assistant_interactions
-                  WHERE meeting_id = ?1 ORDER BY id ASC",
-            )
-            .map_err(map_err)?;
-        let rows = stmt
-            .query_map(rusqlite::params![meeting_id], |r| {
-                let citations_json: String = r.get(2)?;
-                Ok((
-                    r.get::<_, String>(0)?,
-                    r.get::<_, String>(1)?,
-                    citations_json,
-                    r.get::<_, String>(3)?,
-                    r.get::<_, Option<String>>(4)?,
-                    r.get::<_, String>(5)?,
-                ))
-            })
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            let (command, answer, citations_json, status, source_label, created_at) =
-                r.map_err(map_err)?;
-            // A malformed citations blob must never break the read — fall back to an empty list.
-            let citations: Vec<String> = serde_json::from_str(&citations_json).unwrap_or_default();
-            out.push(AssistantInteraction {
-                command,
-                answer,
-                citations,
-                status,
-                source_label,
-                created_at,
-            });
-        }
-        Ok(out)
-    }
-
-    /// Read the persisted @brain THREAD exchanges for `meeting_id` — ONLY rows that carry a
-    /// `thread_id` (legacy voice rows with NULL are EXCLUDED), oldest first — and ONLY when the
-    /// meeting is VISIBLE to the session `unlocked` set. A sealed-and-not-session-unlocked meeting
-    /// returns an EMPTY list, never an error (existence must not leak) — the same gate as
-    /// `list_assistant_interactions_visible`. On seal the rows are PURGED anyway
-    /// (`purge_assistant_interactions_tx`), so the gate is defense-in-depth.
-    pub fn list_assistant_threads_visible(
-        &self,
-        meeting_id: &str,
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<AssistantThreadRow>> {
-        if !self.meeting_is_visible(meeting_id, unlocked)? {
-            return Ok(Vec::new());
-        }
-        let conn = self.lock();
-        let mut stmt = conn
-            .prepare(
-                "SELECT thread_id, anchor_text, command, answer, citations, status, created_at
-                   FROM assistant_interactions
-                  WHERE meeting_id = ?1 AND thread_id IS NOT NULL
-                  ORDER BY id ASC",
-            )
-            .map_err(map_err)?;
-        let rows = stmt
-            .query_map(rusqlite::params![meeting_id], |r| {
-                Ok((
-                    r.get::<_, String>(0)?,
-                    r.get::<_, Option<String>>(1)?,
-                    r.get::<_, String>(2)?,
-                    r.get::<_, String>(3)?,
-                    r.get::<_, String>(4)?,
-                    r.get::<_, String>(5)?,
-                    r.get::<_, String>(6)?,
-                ))
-            })
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            let (thread_id, anchor_text, command, answer, citations_json, status, created_at) =
-                r.map_err(map_err)?;
-            // A malformed citations blob must never break the read — fall back to an empty list.
-            let citations: Vec<String> = serde_json::from_str(&citations_json).unwrap_or_default();
-            out.push(AssistantThreadRow {
-                thread_id,
-                anchor_text,
-                command,
-                answer,
-                citations,
-                status,
-                created_at,
-            });
-        }
-        Ok(out)
-    }
-
     /// LOCK-SAFETY: delete every `assistant_interactions` row for `meeting_ids` within an EXISTING
     /// transaction, so the purge lands in the SAME atomic unit as the plaintext blanking on a seal
     /// (and on the startup reconcile). The Q&A log is plaintext-derived convenience data that mirrors
@@ -9064,1268 +8909,7 @@ impl Db {
         Ok(out)
     }
 
-    // ── self-assembling graph (entities + mentions) ───────────────────────────
-    //
-    // Sink A of the dual-sink: the encrypted DB is the source of truth for the in-app
-    // graph. EVERY read below pushes the same `unlocked` set through `visibility_clause`
-    // over `entity_mentions → meetings → notes n LEFT JOIN folders f`, replicating the
-    // `EXISTS(visible note) OR NOT EXISTS(any note)` predicate of `list_meetings_visible`
-    // verbatim — so a sealed-and-not-unlocked meeting contributes ZERO to nodes, edges,
-    // and counts. The rows persist through sealing; they merely become invisible at read.
 
-    /// Upsert an entity by `(name_ci, kind)`, case-insensitively de-duplicated. Keeps the
-    /// FIRST-SEEN casing in `name` (a later "anna kowalska" does NOT overwrite "Anna Kowalska").
-    /// `name_ci` uses full-Unicode `to_lowercase()` (NOT ASCII-only folding) so accented names
-    /// dedup consistently. Returns the (new or existing) entity id. Race-safe: `INSERT OR IGNORE`
-    /// then re-read, so a concurrent insert resolves to the single winning row.
-    pub fn upsert_entity(&self, name: &str, kind: EntityKind) -> Result<String> {
-        let conn = self.lock();
-        let name = name.trim();
-        let name_ci = name.to_lowercase();
-        let kind_str = kind.as_str();
-        let id = uuid::Uuid::new_v4().to_string();
-        let created_at = chrono::Utc::now().to_rfc3339();
-        // INSERT OR IGNORE: if `(name_ci, kind)` already exists the insert is a no-op (the
-        // existing first-seen casing + id are kept). Either way we re-read the canonical id.
-        conn.execute(
-            "INSERT OR IGNORE INTO entities (id, name, name_ci, kind, created_at)
-             VALUES (?1, ?2, ?3, ?4, ?5)",
-            rusqlite::params![id, name, name_ci, kind_str, created_at],
-        )
-        .map_err(map_err)?;
-        let resolved: String = conn
-            .query_row(
-                "SELECT id FROM entities WHERE name_ci = ?1 AND kind = ?2",
-                rusqlite::params![name_ci, kind_str],
-                |r| r.get(0),
-            )
-            .map_err(map_err)?;
-        Ok(resolved)
-    }
-
-    /// Record that `entity_id` was mentioned in `meeting_id`. Idempotent via the PK
-    /// `(entity_id, meeting_id)` — re-summarize / re-extract never double-counts.
-    pub fn add_mention(&self, entity_id: &str, meeting_id: &str) -> Result<()> {
-        let conn = self.lock();
-        let created_at = chrono::Utc::now().to_rfc3339();
-        conn.execute(
-            "INSERT OR IGNORE INTO entity_mentions (entity_id, meeting_id, created_at)
-             VALUES (?1, ?2, ?3)",
-            rusqlite::params![entity_id, meeting_id, created_at],
-        )
-        .map_err(map_err)?;
-        Ok(())
-    }
-
-    /// All entities that have ≥1 VISIBLE mention, with that visible mention count. An entity
-    /// mentioned ONLY in sealed-and-not-unlocked meetings has count 0 → dropped by `HAVING`,
-    /// so its name (which lived only in encrypted markdown) never reaches the renderer.
-    pub fn list_entities_visible(&self, unlocked: &HashSet<String>) -> Result<Vec<GraphNode>> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        // 2026-07-13 perf audit (MODERATE): this fed BOTH get_graph's node list and list_people
-        // with NO limit, unlike the sibling graph_edges_visible (MAX_GRAPH_EDGES=600, added for
-        // the same reason — see its own comment). Visibility is already enforced in WHERE; a
-        // trailing LIMIT trims magnitude only, ordered by mention count so the most-relevant
-        // entities/people survive the cut on a vault with many.
-        const MAX_VISIBLE_ENTITIES: usize = 500;
-        let sql = format!(
-            "SELECT e.id, e.name, e.kind, COUNT(em.meeting_id) AS cnt
-               FROM entities e
-               JOIN entity_mentions em ON em.entity_id = e.id
-               JOIN meetings m ON m.id = em.meeting_id
-              WHERE (
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id)
-                   OR EXISTS (
-                        SELECT 1 FROM notes n
-                         LEFT JOIN folders f ON f.id = n.folder_id
-                         WHERE n.meeting_id = m.id AND {visible}
-                      )
-                    )
-              GROUP BY e.id, e.name, e.kind
-             HAVING cnt > 0
-              ORDER BY cnt DESC, e.name COLLATE NOCASE ASC
-              LIMIT {MAX_VISIBLE_ENTITIES}"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt
-            .query_map([], |r| {
-                let id: String = r.get(0)?;
-                let name: String = r.get(1)?;
-                let kind_str: String = r.get(2)?;
-                let mention_count: i64 = r.get(3)?;
-                Ok((id, name, kind_str, mention_count))
-            })
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            let (id, name, kind_str, mention_count) = r.map_err(map_err)?;
-            out.push(GraphNode {
-                id,
-                name,
-                kind: EntityKind::from_str(&kind_str)?,
-                mention_count,
-            });
-        }
-        Ok(out)
-    }
-
-    /// The TRUE count of entities with ≥1 VISIBLE mention — same predicate as
-    /// `list_entities_visible` but with NO `LIMIT`, so callers can detect the 500-row cap
-    /// truncating the result and disclose it (added 2026-07-13: the cap trimmed `get_graph`'s
-    /// node list and `list_people`'s roster silently, with no signal the FE could show — the
-    /// existing `hasHidden`/`has_hidden_folders` flag only reports LOCKED folders, so on a vault
-    /// with >500 visible entities and zero locked folders it stayed false while 100+ entities were
-    /// dropped). `list_entities_visible(unlocked).len() < count_entities_visible(unlocked, None)`
-    /// means the cap trimmed rows. `kind` optionally narrows to one `EntityKind` (e.g. `Person`,
-    /// mirroring how `list_people` filters `list_entities_visible`'s output to persons) — the cap
-    /// applies BEFORE that filter, so the all-kinds and Person-only totals must be counted
-    /// separately rather than derived from each other.
-    pub fn count_entities_visible(
-        &self,
-        unlocked: &HashSet<String>,
-        kind: Option<EntityKind>,
-    ) -> Result<i64> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        let kind_filter = match kind {
-            Some(k) => format!("AND e.kind = '{}'", k.as_str()),
-            None => String::new(),
-        };
-        let sql = format!(
-            "SELECT COUNT(*) FROM (
-               SELECT e.id
-                 FROM entities e
-                 JOIN entity_mentions em ON em.entity_id = e.id
-                 JOIN meetings m ON m.id = em.meeting_id
-                WHERE (
-                        NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id)
-                     OR EXISTS (
-                          SELECT 1 FROM notes n
-                           LEFT JOIN folders f ON f.id = n.folder_id
-                           WHERE n.meeting_id = m.id AND {visible}
-                        )
-                      )
-                      {kind_filter}
-                GROUP BY e.id
-               HAVING COUNT(em.meeting_id) > 0
-             )"
-        );
-        conn.query_row(&sql, [], |r| r.get(0)).map_err(map_err)
-    }
-
-    /// The VISIBLE meetings mentioning `entity_id`, newest first, as `VaultSource` chips
-    /// (the same shape the FE uses for backlink chips). Sealed-not-unlocked meetings excluded.
-    pub fn entity_mentions_visible(
-        &self,
-        entity_id: &str,
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<VaultSource>> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        let sql = format!(
-            "SELECT m.id, m.title, m.started_at
-               FROM entity_mentions em
-               JOIN meetings m ON m.id = em.meeting_id
-              WHERE em.entity_id = ?1
-                AND (
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id)
-                   OR EXISTS (
-                        SELECT 1 FROM notes n
-                         LEFT JOIN folders f ON f.id = n.folder_id
-                         WHERE n.meeting_id = m.id AND {visible}
-                      )
-                    )
-              ORDER BY m.started_at DESC, m.id DESC"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt
-            .query_map(rusqlite::params![entity_id], |r| {
-                let meeting_id: String = r.get(0)?;
-                let title: Option<String> = r.get(1)?;
-                let started_at: String = r.get(2)?;
-                Ok(VaultSource {
-                    meeting_id,
-                    title: title.unwrap_or_default(),
-                    started_at,
-                    origin: None,
-                })
-            })
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
-
-    /// Entity↔entity co-occurrence edges: two entities sharing the SAME visible meeting, weighted
-    /// by the number of shared visible meetings. Pair-deduped via `a.entity_id < b.entity_id`
-    /// → exactly one undirected edge per pair, `source < target`. Both endpoints' meetings are
-    /// gated by the visibility predicate, so a co-occurrence in a sealed meeting yields no edge.
-    pub fn graph_edges_visible(&self, unlocked: &HashSet<String>) -> Result<Vec<GraphEdge>> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        // F5: bound the quadratic co-occurrence self-join — return only the strongest edges. The
-        // graph UI (brain-map, ≤60 nodes) consumes only the heaviest connections, so an unbounded
-        // ORDER BY over every co-mentioned entity pair is wasted serialization on a large vault.
-        // Visibility is already enforced in WHERE; a trailing LIMIT trims magnitude only — it can
-        // never widen what is visible.
-        const MAX_GRAPH_EDGES: usize = 600;
-        // PR-9 F3: `weight DESC` alone leaves ties in engine-arbitrary order → the surviving 600-edge
-        // subset could vary between opens, contradicting the graph's "Deterministic" claim. Break ties
-        // on the pair identity (`a.entity_id`, then `b.entity_id`, both already `a < b`-canonicalized
-        // above) so identical data → the identical edge set every call.
-        let sql = format!(
-            "SELECT a.entity_id, b.entity_id, COUNT(*) AS weight
-               FROM entity_mentions a
-               JOIN entity_mentions b
-                 ON a.meeting_id = b.meeting_id AND a.entity_id < b.entity_id
-               JOIN meetings m ON m.id = a.meeting_id
-              WHERE (
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id)
-                   OR EXISTS (
-                        SELECT 1 FROM notes n
-                         LEFT JOIN folders f ON f.id = n.folder_id
-                         WHERE n.meeting_id = m.id AND {visible}
-                      )
-                    )
-              GROUP BY a.entity_id, b.entity_id
-              ORDER BY weight DESC, a.entity_id ASC, b.entity_id ASC
-              LIMIT {MAX_GRAPH_EDGES}"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt
-            .query_map([], |r| {
-                Ok(GraphEdge {
-                    source: r.get(0)?,
-                    target: r.get(1)?,
-                    weight: r.get(2)?,
-                })
-            })
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
-
-    /// Whether `entity_id` has ≥1 VISIBLE mention right now — i.e. at least one mention lands in a
-    /// meeting that is visible under the SAME predicate as `list_meetings_visible` / the other
-    /// graph reads (`EXISTS(visible note) OR NOT EXISTS(any note)`). An entity mentioned ONLY in
-    /// sealed-and-not-unlocked meetings returns `false`, so its name (which lived only in encrypted
-    /// markdown) can never leak through `get_entity` / `build_entity_detail`. This is the gate the
-    /// detail path was missing: `get_entity` itself reads the raw `entities` row with no visibility
-    /// predicate, so callers that expose an entity to the FE MUST go through this check first.
-    pub fn entity_is_visible(&self, entity_id: &str, unlocked: &HashSet<String>) -> Result<bool> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        let sql = format!(
-            "SELECT EXISTS (
-                      SELECT 1
-                        FROM entity_mentions em
-                        JOIN meetings m ON m.id = em.meeting_id
-                       WHERE em.entity_id = ?1
-                         AND (
-                               NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id)
-                            OR EXISTS (
-                                 SELECT 1 FROM notes n
-                                  LEFT JOIN folders f ON f.id = n.folder_id
-                                  WHERE n.meeting_id = m.id AND {visible}
-                               )
-                             )
-                    )"
-        );
-        let visible: bool = conn
-            .query_row(&sql, rusqlite::params![entity_id], |r| {
-                Ok(r.get::<_, i64>(0)? != 0)
-            })
-            .map_err(map_err)?;
-        Ok(visible)
-    }
-
-    /// One entity row by id (`None` if absent), with its first-seen casing. NOTE: this reads the
-    /// raw `entities` row WITHOUT a visibility predicate — it must NOT be exposed to the FE for an
-    /// arbitrary id without first gating on [`entity_is_visible`] (see `build_entity_detail`).
-    pub fn get_entity(&self, entity_id: &str) -> Result<Option<GraphEntity>> {
-        let conn = self.lock();
-        let row = conn
-            .query_row(
-                "SELECT id, name, kind, created_at FROM entities WHERE id = ?1",
-                rusqlite::params![entity_id],
-                |r| {
-                    let id: String = r.get(0)?;
-                    let name: String = r.get(1)?;
-                    let kind_str: String = r.get(2)?;
-                    let created_at: String = r.get(3)?;
-                    Ok((id, name, kind_str, created_at))
-                },
-            )
-            .optional()
-            .map_err(map_err)?;
-        match row {
-            Some((id, name, kind_str, created_at)) => Ok(Some(GraphEntity {
-                id,
-                name,
-                kind: EntityKind::from_str(&kind_str)?,
-                created_at,
-            })),
-            None => Ok(None),
-        }
-    }
-
-    /// The top-`limit` entities co-occurring with `entity_id` (the neighborhood satellites),
-    /// ranked by shared VISIBLE meeting count. Both the anchor's and the neighbor's mention must
-    /// land in a visible meeting, so sealed co-occurrences never surface a neighbor.
-    pub fn entity_neighbors_visible(
-        &self,
-        entity_id: &str,
-        unlocked: &HashSet<String>,
-        limit: i64,
-    ) -> Result<Vec<EntityNeighbor>> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        let sql = format!(
-            "SELECT e.id, e.name, e.kind, COUNT(*) AS shared
-               FROM entity_mentions a
-               JOIN entity_mentions b ON a.meeting_id = b.meeting_id AND b.entity_id != a.entity_id
-               JOIN entities e ON e.id = b.entity_id
-               JOIN meetings m ON m.id = a.meeting_id
-              WHERE a.entity_id = ?1
-                AND (
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id)
-                   OR EXISTS (
-                        SELECT 1 FROM notes n
-                         LEFT JOIN folders f ON f.id = n.folder_id
-                         WHERE n.meeting_id = m.id AND {visible}
-                      )
-                    )
-              GROUP BY e.id, e.name, e.kind
-              ORDER BY shared DESC, e.name COLLATE NOCASE ASC
-              LIMIT ?2"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt
-            .query_map(rusqlite::params![entity_id, limit], |r| {
-                let id: String = r.get(0)?;
-                let name: String = r.get(1)?;
-                let kind_str: String = r.get(2)?;
-                let shared_meetings: i64 = r.get(3)?;
-                Ok((id, name, kind_str, shared_meetings))
-            })
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            let (id, name, kind_str, shared_meetings) = r.map_err(map_err)?;
-            out.push(EntityNeighbor {
-                id,
-                name,
-                kind: EntityKind::from_str(&kind_str)?,
-                shared_meetings,
-            });
-        }
-        Ok(out)
-    }
-
-    /// QUERY→ENTITY resolution for GraphRAG-lite (Phase 2d) — DETERMINISTIC, no LLM. Returns the
-    /// ids of VISIBLE entities whose name appears as a whole-token match (see
-    /// [`name_matches_query_tokens`]) inside `query`, case-insensitively. Gated by EXACTLY the
-    /// `list_entities_visible` predicate: an entity mentioned ONLY in a sealed-and-not-unlocked
-    /// folder is never resolved (its name lived only in encrypted markdown, so resolving it would
-    /// leak its existence). A name shorter than [`MIN_ENTITY_NAME_LEN`] chars is skipped (noise
-    /// guard). Empty query or no match → empty vec, leaving the hybrid path unchanged.
-    pub fn entities_matching_query(
-        &self,
-        query: &str,
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<String>> {
-        let q_tokens = tokenize_lower(query);
-        if q_tokens.is_empty() {
-            return Ok(Vec::new());
-        }
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        // Same visibility predicate as `list_entities_visible`: keep an entity iff it has ≥1
-        // mention in a meeting that is open/NULL-folder OR session-unlocked (or note-less).
-        let sql = format!(
-            "SELECT DISTINCT e.id, e.name_ci
-               FROM entities e
-               JOIN entity_mentions em ON em.entity_id = e.id
-               JOIN meetings m ON m.id = em.meeting_id
-              WHERE (
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id)
-                   OR EXISTS (
-                        SELECT 1 FROM notes n
-                         LEFT JOIN folders f ON f.id = n.folder_id
-                         WHERE n.meeting_id = m.id AND {visible}
-                      )
-                    )"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt
-            .query_map([], |r| {
-                let id: String = r.get(0)?;
-                let name_ci: String = r.get(1)?;
-                Ok((id, name_ci))
-            })
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            let (id, name_ci) = r.map_err(map_err)?;
-            if name_ci.chars().count() >= MIN_ENTITY_NAME_LEN
-                && name_matches_query_tokens(&q_tokens, &name_ci)
-            {
-                out.push(id);
-            }
-        }
-        Ok(out)
-    }
-
-    /// GATED entity-neighbour candidates for GraphRAG-lite (Phase 2d): the VISIBLE meetings
-    /// mentioning ANY of `entity_ids`, ranked by how many of the matched entities they touch (desc)
-    /// then recency. Uses EXACTLY the `list_meetings_visible`/graph visibility predicate, so a
-    /// sealed-and-not-unlocked meeting NEVER appears even if it mentions a matched entity. Empty
-    /// input → empty vec.
-    pub fn meetings_mentioning_entities_visible(
-        &self,
-        entity_ids: &[String],
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<Meeting>> {
-        if entity_ids.is_empty() {
-            return Ok(Vec::new());
-        }
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        let placeholders = (1..=entity_ids.len())
-            .map(|i| format!("?{i}"))
-            .collect::<Vec<_>>()
-            .join(", ");
-        let sql = format!(
-            "SELECT m.id, m.started_at, m.ended_at, m.title, m.duration_s, m.audio_path, m.status, \
-                    (SELECT nf.folder_id FROM notes nf \
-                      WHERE nf.meeting_id = m.id AND nf.folder_id IS NOT NULL LIMIT 1) AS folder_id \
-               FROM entity_mentions em \
-               JOIN meetings m ON m.id = em.meeting_id \
-              WHERE em.entity_id IN ({placeholders}) \
-                AND ( \
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id) \
-                   OR EXISTS ( \
-                        SELECT 1 FROM notes n \
-                         LEFT JOIN folders f ON f.id = n.folder_id \
-                         WHERE n.meeting_id = m.id AND {visible} \
-                      ) \
-                    ) \
-              GROUP BY m.id \
-              ORDER BY COUNT(DISTINCT em.entity_id) DESC, m.started_at DESC, m.id DESC"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let params = rusqlite::params_from_iter(entity_ids.iter());
-        let rows = stmt.query_map(params, row_to_meeting).map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)??);
-        }
-        Ok(out)
-    }
-
-    /// Whether ANY folder is sealed-and-not-unlocked right now (i.e. a locked folder whose id is
-    /// NOT in the session `unlocked` set). Drives the FE's one honest "some entities hidden"
-    /// disclosure banner — it never leaks how many or which.
-    pub fn has_hidden_folders(&self, unlocked: &HashSet<String>) -> Result<bool> {
-        let conn = self.lock();
-        let mut stmt = conn
-            .prepare("SELECT id FROM folders WHERE locked = 1")
-            .map_err(map_err)?;
-        let rows = stmt
-            .query_map([], |r| r.get::<_, String>(0))
-            .map_err(map_err)?;
-        for r in rows {
-            let id = r.map_err(map_err)?;
-            if !unlocked.contains(&id) {
-                return Ok(true);
-            }
-        }
-        Ok(false)
-    }
-
-    /// Build the full graph payload (`get_graph`): all visible nodes + all visible edges +
-    /// the hidden-folder disclosure flag, snapshotting the passed-in session `unlocked` set.
-    pub fn build_graph(&self, unlocked: &HashSet<String>) -> Result<GraphData> {
-        let nodes = self.list_entities_visible(unlocked)?;
-        let edges = self.graph_edges_visible(unlocked)?;
-        let has_hidden = self.has_hidden_folders(unlocked)?;
-        let total_visible_entities = self.count_entities_visible(unlocked, None)?;
-        Ok(GraphData {
-            nodes,
-            edges,
-            has_hidden,
-            total_visible_entities,
-        })
-    }
-
-    /// Build the FULL-BRAIN graph payload (`get_full_graph`, DESIGN §PR-4). A PURE READ — no writes,
-    /// no new storage — that unifies FOUR node kinds and FIVE edge kinds under one visibility model:
-    ///
-    /// NODES (each via its EXISTING gated reader; a sealed-and-not-session-unlocked item yields none):
-    ///   • entities  — `list_entities_visible` (already 500-capped, HAVING visible mention);
-    ///   • meetings  — `list_meetings_visible` (visible if no notes OR any visible note);
-    ///   • notes + documents — `full_graph_content_nodes` over `documents JOIN folders` under
-    ///     `visibility_clause` (a sealed folder's rows are absent), split by `kind`.
-    ///
-    /// EDGES (BOTH endpoints must be in the visible-node set built above — an edge to a sealed node
-    /// is DROPPED, so no edge can leak a hidden node's existence):
-    ///   • co_occurrence — entity↔entity (`graph_edges_visible`, already gated);
-    ///   • mention       — entity→meeting (`entity_mentions`, gated by the SAME meeting predicate);
-    ///   • wikilink/companion/semantic — `links` rows with `status='active'` (+ `status='suggested'`
-    ///     semantic ONLY when `opts.include_suggested`), each endpoint re-checked against the set.
-    ///
-    /// Every edge carries `src_kind`/`dst_kind` (the endpoint node kinds it was gated on) so the FE
-    /// matches endpoints by `(kind, id)`, safe against a cross-kind id collision (PR-9 F4).
-    ///
-    /// Deterministic: nodes ordered by (kind, id ASC), edges by (kind, src, dst, status); every edge
-    /// leg has a deterministic `ORDER BY` before its cap (co_occurrence by weight then pair id;
-    /// mention by meeting recency; links by score). Honest caps: `total_visible_nodes` is the pre-cap
-    /// true NODE count so the FE can disclose a silent node trim; `edges_truncated` is true when an
-    /// EDGE-leg cap (mention/links LIMIT) trimmed edges; `has_hidden` reflects LOCKED folders (mirrors
-    /// the entity graph). No PII logged — ids/counts only.
-    pub fn build_full_graph(
-        &self,
-        unlocked: &HashSet<String>,
-        opts: FullGraphOpts,
-    ) -> Result<FullGraphData> {
-        // ── NODES: enumerate via the existing gated readers, keyed by (kind_str, id) so a cross-kind
-        //    id collision can never conflate two nodes when an edge is gated. ──
-        let mut nodes: Vec<FullGraphNode> = Vec::new();
-        let mut visible: HashSet<(&'static str, String)> = HashSet::new();
-
-        // Entities (already render-capped at 500 inside the reader).
-        for e in self.list_entities_visible(unlocked)? {
-            visible.insert((FullGraphNodeKind::Entity.as_str(), e.id.clone()));
-            nodes.push(FullGraphNode {
-                id: e.id,
-                kind: FullGraphNodeKind::Entity,
-                label: e.name,
-                date: None,
-                degree: 0,
-            });
-        }
-        // Meetings (visible-meeting predicate; capped to keep the payload bounded).
-        for m in self.list_meetings_visible(MAX_FULL_GRAPH_PER_KIND as i64, unlocked)? {
-            visible.insert((FullGraphNodeKind::Meeting.as_str(), m.id.clone()));
-            let label = m
-                .title
-                .as_deref()
-                .map(str::trim)
-                .filter(|t| !t.is_empty())
-                .unwrap_or("Meeting")
-                .to_string();
-            nodes.push(FullGraphNode {
-                id: m.id,
-                kind: FullGraphNodeKind::Meeting,
-                label,
-                date: Some(m.started_at),
-                degree: 0,
-            });
-        }
-        // Notes + documents (one gated pass over `documents`, split by kind).
-        for (kind, id, label, date) in self.full_graph_content_nodes(unlocked)? {
-            visible.insert((kind.as_str(), id.clone()));
-            nodes.push(FullGraphNode {
-                id,
-                kind,
-                label,
-                date,
-                degree: 0,
-            });
-        }
-        // TRUE (uncapped) visible-node count for honest disclosure — computed with NO per-kind LIMIT,
-        // so `total_visible_nodes > nodes.len()` whenever a per-kind cap silently trimmed a leg (the
-        // twin of the entity graph's `total_visible_entities`). Mirrors, never derives from, the capped
-        // `nodes.len()`.
-        let total_visible_nodes = self.count_full_graph_nodes_visible(unlocked)?;
-
-        // ── EDGES: every leg is dropped unless BOTH endpoints are in `visible`. Degree is tallied
-        //    INLINE with the true endpoint node-kinds (a links edge can connect meeting↔note, so the
-        //    endpoint kinds are NOT derivable from the edge kind alone — they must be carried here). ──
-        let mut edges: Vec<FullGraphEdge> = Vec::new();
-        let mut degree: std::collections::HashMap<(&'static str, String), i64> =
-            std::collections::HashMap::new();
-        let bump = |degree: &mut std::collections::HashMap<(&'static str, String), i64>,
-                    k: &'static str,
-                    id: &str| {
-            *degree.entry((k, id.to_string())).or_insert(0) += 1;
-        };
-        let ent = FullGraphNodeKind::Entity.as_str();
-        let mtg = FullGraphNodeKind::Meeting.as_str();
-
-        // co_occurrence — entity↔entity (already visibility-gated; re-check both endpoints against the
-        // capped node set so an edge to a >500-cap-dropped entity is also dropped).
-        for ge in self.graph_edges_visible(unlocked)? {
-            if visible.contains(&(ent, ge.source.clone())) && visible.contains(&(ent, ge.target.clone()))
-            {
-                bump(&mut degree, ent, &ge.source);
-                bump(&mut degree, ent, &ge.target);
-                edges.push(FullGraphEdge {
-                    src: ge.source,
-                    dst: ge.target,
-                    src_kind: FullGraphNodeKind::Entity,
-                    dst_kind: FullGraphNodeKind::Entity,
-                    kind: FullGraphEdgeKind::CoOccurrence,
-                    score: ge.weight as f64,
-                    status: "active".to_string(),
-                });
-            }
-        }
-        // mention — entity→meeting (gated by the SAME meeting-visible predicate; both endpoints then
-        // re-checked against the node set, so a mention into a sealed OR cap-dropped meeting is gone).
-        // `edges_truncated` accumulates whether ANY edge leg hit its cap, so the FE can disclose it.
-        let (mention_edges, mentions_truncated) = self.entity_meeting_mentions_visible(unlocked)?;
-        let (link_rows, links_truncated) = self.full_graph_links(opts.include_suggested)?;
-        // True when a genuine EDGE-leg cap (the mention or links LIMIT) trimmed rows — distinct from
-        // an edge dropped because its endpoint fell to a NODE cap (that is the node disclosure's job).
-        let edges_truncated = mentions_truncated || links_truncated;
-        for (entity_id, meeting_id, weight) in mention_edges {
-            if visible.contains(&(ent, entity_id.clone())) && visible.contains(&(mtg, meeting_id.clone()))
-            {
-                bump(&mut degree, ent, &entity_id);
-                bump(&mut degree, mtg, &meeting_id);
-                edges.push(FullGraphEdge {
-                    src: entity_id,
-                    dst: meeting_id,
-                    src_kind: FullGraphNodeKind::Entity,
-                    dst_kind: FullGraphNodeKind::Meeting,
-                    kind: FullGraphEdgeKind::Mention,
-                    score: weight as f64,
-                    status: "active".to_string(),
-                });
-            }
-        }
-        // links rows — wikilink/companion/semantic. active always; suggested semantic behind the flag.
-        // BOTH endpoints re-checked against the visible-node set (the links kind strings map 1:1 onto
-        // the node-kind strings: meeting|note|document), so a link to a sealed/absent node is dropped.
-        for (src_kind, src_id, dst_kind, dst_id, edge_type, score, status) in link_rows {
-            let (Some(src_nk), Some(dst_nk)) = (
-                full_graph_link_node_kind(&src_kind),
-                full_graph_link_node_kind(&dst_kind),
-            ) else {
-                continue; // corrupt/unknown kind → skip defensively.
-            };
-            let (src_k, dst_k) = (src_nk.as_str(), dst_nk.as_str());
-            if !visible.contains(&(src_k, src_id.clone()))
-                || !visible.contains(&(dst_k, dst_id.clone()))
-            {
-                continue;
-            }
-            let Some(kind) = full_graph_edge_kind_from_type(&edge_type) else {
-                continue; // unknown edge_type → skip.
-            };
-            bump(&mut degree, src_k, &src_id);
-            bump(&mut degree, dst_k, &dst_id);
-            edges.push(FullGraphEdge {
-                src: src_id,
-                dst: dst_id,
-                src_kind: src_nk,
-                dst_kind: dst_nk,
-                kind,
-                score,
-                status,
-            });
-        }
-
-        // ── Degrees: assign each node its in-graph incident-edge count (a layout hint, never the true
-        //    corpus degree). Keyed by (kind, id) to avoid cross-kind id collisions. ──
-        for n in &mut nodes {
-            n.degree = degree
-                .get(&(n.kind.as_str(), n.id.clone()))
-                .copied()
-                .unwrap_or(0);
-        }
-
-        // Deterministic ordering: nodes by (kind, id ASC); edges by (kind, src, dst, status).
-        nodes.sort_by(|a, b| {
-            a.kind
-                .as_str()
-                .cmp(b.kind.as_str())
-                .then_with(|| a.id.cmp(&b.id))
-        });
-        edges.sort_by(|a, b| {
-            a.kind
-                .as_str()
-                .cmp(b.kind.as_str())
-                .then_with(|| a.src.cmp(&b.src))
-                .then_with(|| a.dst.cmp(&b.dst))
-                .then_with(|| a.status.cmp(&b.status))
-        });
-
-        let has_hidden = self.has_hidden_folders(unlocked)?;
-        tracing::debug!(
-            target: "graph",
-            nodes = nodes.len(),
-            edges = edges.len(),
-            has_hidden,
-            edges_truncated,
-            "build_full_graph resolved"
-        );
-        Ok(FullGraphData {
-            nodes,
-            edges,
-            has_hidden,
-            total_visible_nodes,
-            edges_truncated,
-        })
-    }
-
-    /// The VISIBLE note + document nodes for the full-brain graph: `(kind, id, label, date)` per row,
-    /// gated by `visibility_clause` over `documents JOIN folders` (a sealed-and-not-session-unlocked
-    /// folder's rows are ABSENT — never masked). `kind='note'` → [`FullGraphNodeKind::Note`], anything
-    /// else → [`FullGraphNodeKind::Document`]. `label` = title, else the filesystem `name`. `date` =
-    /// the `updated_at`/`created_at` timestamp UNIFIED to an RFC3339 ISO string (PR-9 F4 — meetings
-    /// already emit ISO `started_at`; notes/docs previously emitted the raw epoch-ms as a string,
-    /// giving the FE two incompatible date formats on the SAME field). Per-kind capped
-    /// (`MAX_FULL_GRAPH_PER_KIND`) ordered newest-first so the most-relevant rows survive on a big
-    /// vault; visibility is enforced in WHERE, so the cap trims magnitude only.
-    pub(crate) fn full_graph_content_nodes(
-        &self,
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<FullGraphContentNode>> {
-        let conn = self.lock();
-        let visible = visibility_clause("f", unlocked);
-        // Two capped passes (one per kind) so notes and documents each get their own budget rather
-        // than one starving the other on a lopsided vault.
-        let mut out: Vec<FullGraphContentNode> = Vec::new();
-        for (kind_str, node_kind) in [
-            ("note", FullGraphNodeKind::Note),
-            ("document", FullGraphNodeKind::Document),
-        ] {
-            let sql = format!(
-                "SELECT d.id, COALESCE(NULLIF(TRIM(d.title), ''), d.name),
-                        COALESCE(d.updated_at, d.created_at)
-                   FROM documents d
-                   JOIN folders f ON f.id = d.folder_id
-                  WHERE d.kind = ?1 AND {visible}
-                  ORDER BY COALESCE(d.updated_at, d.created_at) DESC, d.id ASC
-                  LIMIT {MAX_FULL_GRAPH_PER_KIND}"
-            );
-            let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-            let rows = stmt
-                .query_map(rusqlite::params![kind_str], |r| {
-                    let id: String = r.get(0)?;
-                    let label: String = r.get(1)?;
-                    let ts: Option<i64> = r.get(2)?;
-                    Ok((id, label, ts))
-                })
-                .map_err(map_err)?;
-            for r in rows {
-                let (id, label, ts) = r.map_err(map_err)?;
-                // Unify to RFC3339 so the FE sees ONE date format across every node kind. epoch-ms →
-                // UTC ISO; an out-of-range/absent timestamp yields None (a bad hint, never a panic).
-                let date = ts.and_then(epoch_ms_to_rfc3339);
-                out.push((node_kind, id, label, date));
-            }
-        }
-        Ok(out)
-    }
-
-    /// The TRUE (UNCAPPED) count of VISIBLE full-brain nodes: entities + meetings + notes +
-    /// documents, each under the SAME visibility predicate as its node reader but with NO per-kind
-    /// LIMIT. Drives `FullGraphData::total_visible_nodes` so the FE can disclose when a per-kind
-    /// render cap silently trimmed the graph (mirrors `count_entities_visible` for the entity graph).
-    /// A sealed-and-not-session-unlocked item is NOT counted (leak-free — a bare total).
-    fn count_full_graph_nodes_visible(&self, unlocked: &HashSet<String>) -> Result<i64> {
-        // Entities reuse the dedicated uncapped counter (same HAVING-visible-mention predicate).
-        let entities = self.count_entities_visible(unlocked, None)?;
-        let conn = self.lock();
-        // Meetings — the `list_meetings_visible` predicate, uncapped.
-        let m_visible = visibility_clause("f", unlocked);
-        let meetings: i64 = conn
-            .query_row(
-                &format!(
-                    "SELECT COUNT(*) FROM meetings m
-                       WHERE NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id)
-                          OR EXISTS (SELECT 1 FROM notes n
-                                       LEFT JOIN folders f ON f.id = n.folder_id
-                                      WHERE n.meeting_id = m.id AND {m_visible})"
-                ),
-                [],
-                |r| r.get(0),
-            )
-            .map_err(map_err)?;
-        // Notes + documents — folder-gated, uncapped, both kinds.
-        let d_visible = visibility_clause("f", unlocked);
-        let docs: i64 = conn
-            .query_row(
-                &format!(
-                    "SELECT COUNT(*) FROM documents d
-                       JOIN folders f ON f.id = d.folder_id
-                      WHERE d.kind IN ('note', 'document') AND {d_visible}"
-                ),
-                [],
-                |r| r.get(0),
-            )
-            .map_err(map_err)?;
-        Ok(entities + meetings + docs)
-    }
-
-    /// Entity→meeting MENTION edges for the full-brain graph: `(entity_id, meeting_id, weight)` for
-    /// every mention landing in a VISIBLE meeting (the SAME `EXISTS(visible note) OR NOT EXISTS(any
-    /// note)` predicate as `list_meetings_visible`/`graph_edges_visible`). `weight` is always 1 (one
-    /// mention row per (entity, meeting) — the PK guarantees it); kept as a field for edge-uniformity.
-    /// A mention into a sealed-and-not-session-unlocked meeting yields NO row, so it can never surface
-    /// a hidden meeting.
-    ///
-    /// PR-9 F2: the previous `ORDER BY em.entity_id ASC` truncated at `MAX_MENTION_EDGES` in
-    /// arbitrary entity-UUID order, so entities LATE in UUID order silently lost ALL their mention
-    /// edges (rendered isolated, degree 0, with no signal). Order by meeting RECENCY first
-    /// (`m.started_at DESC`) so the freshest mentions survive the cap regardless of UUID, with a
-    /// deterministic `(entity_id, meeting_id)` tiebreak so identical data → the identical subset.
-    /// Returns `(rows, truncated)` — `truncated` is true when the cap trimmed rows (there was a
-    /// `MAX_MENTION_EDGES + 1`th visible mention), so the caller can disclose the edge trim.
-    fn entity_meeting_mentions_visible(
-        &self,
-        unlocked: &HashSet<String>,
-    ) -> Result<(Vec<FullGraphMentionEdge>, bool)> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        const MAX_MENTION_EDGES: usize = 2000;
-        // Fetch ONE past the cap so a full page distinguishes "exactly the cap" from "trimmed".
-        let sql = format!(
-            "SELECT em.entity_id, em.meeting_id
-               FROM entity_mentions em
-               JOIN meetings m ON m.id = em.meeting_id
-              WHERE (
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id)
-                   OR EXISTS (
-                        SELECT 1 FROM notes n
-                         LEFT JOIN folders f ON f.id = n.folder_id
-                         WHERE n.meeting_id = m.id AND {visible}
-                      )
-                    )
-              ORDER BY m.started_at DESC, em.entity_id ASC, em.meeting_id ASC
-              LIMIT {}",
-            MAX_MENTION_EDGES + 1
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt
-            .query_map([], |r| {
-                Ok((r.get::<_, String>(0)?, r.get::<_, String>(1)?, 1i64))
-            })
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        let truncated = out.len() > MAX_MENTION_EDGES;
-        out.truncate(MAX_MENTION_EDGES);
-        Ok((out, truncated))
-    }
-
-    /// Raw `links` rows for the full-brain graph, PRE-gating (the caller gates both endpoints against
-    /// the visible-node set). `status='active'` always; `status='suggested'` semantic rows ONLY when
-    /// `include_suggested`. `dismissed` rows are NEVER returned. Returns
-    /// `(rows, truncated)` where each row is `(src_kind, src_id, dst_kind, dst_id, edge_type, score,
-    /// status)` ordered deterministically. This reads no content columns — ids/kinds/metadata only —
-    /// and is gated by the caller, so it is leak-free at every call site.
-    ///
-    /// PR-9 F2: `links` is the FASTEST-growing edge leg (every wikilink/companion/semantic suggestion
-    /// is a row) and was read UNBOUNDED — the per-kind node caps existed while the highest-cardinality
-    /// edge table had none. Bound it at `MAX_FULL_GRAPH_LINK_EDGES`, ordered by score DESC (strongest
-    /// links survive) with a deterministic `(edge_type, src_id, dst_id, id)` tiebreak, and report
-    /// `truncated` so the caller can disclose the edge trim (mirrors the mention-edge cap).
-    fn full_graph_links(&self, include_suggested: bool) -> Result<(Vec<FullGraphLinkRow>, bool)> {
-        let conn = self.lock();
-        // active: any edge_type. suggested: ONLY semantic (wikilink/companion are always active by
-        // construction; there is no "suggested wikilink"). dismissed tombstones stay excluded.
-        let status_pred = if include_suggested {
-            "(status = 'active' OR (status = 'suggested' AND edge_type = 'semantic'))"
-        } else {
-            "status = 'active'"
-        };
-        // Fetch ONE past the cap to distinguish "exactly the cap" from "trimmed".
-        let sql = format!(
-            "SELECT src_kind, src_id, dst_kind, dst_id, edge_type, score, status
-               FROM links
-              WHERE {status_pred}
-              ORDER BY score DESC, edge_type ASC, src_id ASC, dst_id ASC, id ASC
-              LIMIT {}",
-            MAX_FULL_GRAPH_LINK_EDGES + 1
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt
-            .query_map([], |r| {
-                Ok((
-                    r.get::<_, String>(0)?,
-                    r.get::<_, String>(1)?,
-                    r.get::<_, String>(2)?,
-                    r.get::<_, String>(3)?,
-                    r.get::<_, String>(4)?,
-                    r.get::<_, f64>(5)?,
-                    r.get::<_, String>(6)?,
-                ))
-            })
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        let truncated = out.len() > MAX_FULL_GRAPH_LINK_EDGES;
-        out.truncate(MAX_FULL_GRAPH_LINK_EDGES);
-        Ok((out, truncated))
-    }
-
-    /// Build the detail payload for one entity (`get_entity_detail`): the entity, its visible
-    /// backlinked meetings, and its top co-occurring neighbors. `None` if the entity is unknown
-    /// OR has ZERO visible mentions (mentioned only in sealed-not-unlocked meetings). The
-    /// visibility gate is mandatory here: `get_entity` reads the raw `entities` row with NO
-    /// predicate, so without this check a caller holding a stale entity id (cached from a prior
-    /// open-folder `get_graph`, before the folder was sealed/auto-relocked) could read back the
-    /// entity's `name` — which lived only in the sealed meeting's encrypted markdown. Routing
-    /// through `entity_is_visible` keeps the detail path consistent with every other graph read.
-    pub fn build_entity_detail(
-        &self,
-        entity_id: &str,
-        unlocked: &HashSet<String>,
-        neighbor_limit: i64,
-    ) -> Result<Option<EntityDetail>> {
-        // Anti-leak gate FIRST: a sealed-only entity is indistinguishable from an unknown id.
-        if !self.entity_is_visible(entity_id, unlocked)? {
-            return Ok(None);
-        }
-        let entity = match self.get_entity(entity_id)? {
-            Some(e) => e,
-            None => return Ok(None),
-        };
-        let meetings = self.entity_mentions_visible(entity_id, unlocked)?;
-        let neighbors = self.entity_neighbors_visible(entity_id, unlocked, neighbor_limit)?;
-        Ok(Some(EntityDetail {
-            entity,
-            meetings,
-            neighbors,
-        }))
-    }
-
-    // ── bitemporal FACTS layer (brain2 R2) ────────────────────────────────────
-    //
-    // Facts are DERIVED content tied to a meeting (the `meeting_id` anchor). The reconcile engine
-    // (`crate::facts`) is pure + deterministic; these methods are the persistence + the GATED read.
-    // LOCK MODEL: `facts_for_entities` is an INTERNAL un-gated read used ONLY by the pipeline to
-    // reconcile (never exposed to the FE — like `raw_segments`); every USER-FACING read goes through
-    // `list_facts_visible`, and a sealed meeting's facts are PURGED on seal (`purge_facts_tx`).
-
-    /// ALL facts (open + closed) for `entity_ids` — the reconcile input. INTERNAL: this is the
-    /// un-gated lifecycle read (the pipeline reconciles before any seal can hide rows), NOT a
-    /// user-facing surface. Empty input → empty vec.
-    pub fn facts_for_entities(&self, entity_ids: &[String]) -> Result<Vec<crate::facts::Fact>> {
-        if entity_ids.is_empty() {
-            return Ok(Vec::new());
-        }
-        let conn = self.lock();
-        let placeholders = (1..=entity_ids.len())
-            .map(|i| format!("?{i}"))
-            .collect::<Vec<_>>()
-            .join(", ");
-        let sql = format!(
-            "SELECT id, entity_id, subject, predicate, object, valid_from, valid_to, recorded_at, \
-                    meeting_id, confidence \
-               FROM facts WHERE entity_id IN ({placeholders}) ORDER BY recorded_at ASC, id ASC"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let params = rusqlite::params_from_iter(entity_ids.iter());
-        let rows = stmt.query_map(params, row_to_fact).map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
-
-    /// Apply a batch of reconcile [`FactOp`]s in ONE atomic transaction: INSERT each `Add`, set
-    /// `valid_to` on each `Invalidate` (only if still open — idempotent), skip `NoOp`. A fresh UUID
-    /// is minted per Add. The whole batch commits or rolls back together, so a crash mid-apply never
-    /// leaves a half-reconciled (e.g. old closed but new not added) store.
-    pub fn apply_fact_ops(&self, ops: &[crate::facts::FactOp]) -> Result<()> {
-        use crate::facts::FactOp;
-        if ops.is_empty() {
-            return Ok(());
-        }
-        let mut conn = self.lock();
-        let tx = conn.transaction().map_err(map_err)?;
-        for op in ops {
-            match op {
-                FactOp::Add(nf) => {
-                    let id = uuid::Uuid::new_v4().to_string();
-                    tx.execute(
-                        "INSERT INTO facts \
-                           (id, entity_id, subject, predicate, object, valid_from, valid_to, \
-                            recorded_at, meeting_id, confidence) \
-                         VALUES (?1, ?2, ?3, ?4, ?5, ?6, NULL, ?7, ?8, ?9)",
-                        rusqlite::params![
-                            id,
-                            nf.entity_id,
-                            nf.subject,
-                            nf.predicate,
-                            nf.object,
-                            nf.valid_from,
-                            nf.recorded_at,
-                            nf.meeting_id,
-                            nf.confidence,
-                        ],
-                    )
-                    .map_err(map_err)?;
-                }
-                FactOp::Invalidate { id, valid_to } => {
-                    tx.execute(
-                        "UPDATE facts SET valid_to = ?2 WHERE id = ?1 AND valid_to IS NULL",
-                        rusqlite::params![id, valid_to],
-                    )
-                    .map_err(map_err)?;
-                }
-                FactOp::NoOp => {}
-            }
-        }
-        tx.commit().map_err(map_err)?;
-        Ok(())
-    }
-
-    /// GATED read: the VISIBLE facts for `entity_id` (open + recently-closed), newest valid_from
-    /// first. A fact is visible iff its source meeting is visible under the SAME predicate as every
-    /// other graph/MCP read (`EXISTS(visible note) OR NOT EXISTS(any note)`), so a
-    /// sealed-and-not-session-unlocked meeting's facts surface NOTHING. A fact with a NULL
-    /// `meeting_id` (legacy/unattributed) is NOT visible — the INNER JOIN to `meetings` drops it
-    /// (fail-closed). This is the single user-facing fact read (UI dossier + egress-free MCP).
-    pub fn list_facts_visible(
-        &self,
-        entity_id: &str,
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<crate::facts::Fact>> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        let sql = format!(
-            "SELECT ft.id, ft.entity_id, ft.subject, ft.predicate, ft.object, ft.valid_from, \
-                    ft.valid_to, ft.recorded_at, ft.meeting_id, ft.confidence \
-               FROM facts ft \
-               JOIN meetings m ON m.id = ft.meeting_id \
-              WHERE ft.entity_id = ?1 \
-                AND ( \
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id) \
-                   OR EXISTS ( \
-                        SELECT 1 FROM notes n \
-                         LEFT JOIN folders f ON f.id = n.folder_id \
-                         WHERE n.meeting_id = m.id AND {visible} \
-                      ) \
-                    ) \
-              ORDER BY (ft.valid_to IS NULL) DESC, ft.valid_from DESC, ft.id DESC"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt
-            .query_map(rusqlite::params![entity_id], row_to_fact)
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
-
-    // ── Re-Truth (supersessions) ────────────────────────────────────────────────
-    //
-    // These are RAW lifecycle reads/writes of the `supersessions` table; the CONTENT gate
-    // (folder-lock + `meeting_is_unlocked`) lives in the commands, exactly like `facts_for_entities`
-    // (raw) vs `list_facts_visible` (gated). Rows are only ever RECORDED for open-folder sources, and
-    // the read command re-gates, so a raw list never leaks a sealed source note's bytes.
-
-    /// Record supersession rows, skipping any whose natural key already exists (idempotent across
-    /// re-summarize — a plain re-summarize reconciles to `NoOp` and emits no new `Invalidate`, but
-    /// this dedupe is the belt-and-suspenders guard). Returns how many NEW rows were inserted. Each
-    /// row is inserted UNAPPLIED (`applied_at` + both pre-images NULL); the pre-images are filled only
-    /// at apply time. One atomic transaction for the batch.
-    pub fn record_supersessions(
-        &self,
-        rows: &[crate::storage::models::SupersessionRow],
-    ) -> Result<usize> {
-        if rows.is_empty() {
-            return Ok(0);
-        }
-        let mut conn = self.lock();
-        let tx = conn.transaction().map_err(map_err)?;
-        let mut inserted = 0usize;
-        for r in rows {
-            // Natural key = the full assertion (both meetings + entity/predicate/old/new). A duplicate
-            // is a no-op regardless of applied state, so a re-record never resurrects a stamped row.
-            let exists: i64 = tx
-                .query_row(
-                    "SELECT COUNT(*) FROM supersessions
-                       WHERE superseding_meeting_id = ?1 AND source_meeting_id = ?2
-                         AND entity = ?3 AND predicate = ?4 AND old_value = ?5 AND new_value = ?6",
-                    rusqlite::params![
-                        r.superseding_meeting_id,
-                        r.source_meeting_id,
-                        r.entity,
-                        r.predicate,
-                        r.old_value,
-                        r.new_value,
-                    ],
-                    |row| row.get(0),
-                )
-                .map_err(map_err)?;
-            if exists > 0 {
-                continue;
-            }
-            tx.execute(
-                "INSERT INTO supersessions
-                   (id, superseding_meeting_id, source_meeting_id, entity, predicate,
-                    old_value, new_value, created_at, applied_at, source_pre_image,
-                    superseding_pre_image)
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, NULL, NULL, NULL)",
-                rusqlite::params![
-                    r.id,
-                    r.superseding_meeting_id,
-                    r.source_meeting_id,
-                    r.entity,
-                    r.predicate,
-                    r.old_value,
-                    r.new_value,
-                    r.created_at,
-                ],
-            )
-            .map_err(map_err)?;
-            inserted += 1;
-        }
-        tx.commit().map_err(map_err)?;
-        Ok(inserted)
-    }
-
-    /// RAW read: UNAPPLIED supersession rows for one superseding meeting, oldest first. The command
-    /// re-gates each row on its SOURCE meeting (folder-lock + unlock) before surfacing it.
-    pub fn unapplied_supersessions_for(
-        &self,
-        superseding_meeting_id: &str,
-    ) -> Result<Vec<crate::storage::models::SupersessionRow>> {
-        let conn = self.lock();
-        let mut stmt = conn
-            .prepare(
-                "SELECT id, superseding_meeting_id, source_meeting_id, entity, predicate,
-                        old_value, new_value, created_at, applied_at, source_pre_image,
-                        superseding_pre_image
-                   FROM supersessions
-                  WHERE superseding_meeting_id = ?1 AND applied_at IS NULL
-                  ORDER BY created_at ASC, id ASC",
-            )
-            .map_err(map_err)?;
-        let rows = stmt
-            .query_map(
-                rusqlite::params![superseding_meeting_id],
-                row_to_supersession,
-            )
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
-
-    /// RAW read: EVERY supersession row that touches `meeting_id` on either side (source OR
-    /// superseding), oldest first. Used by partial-undo to enumerate the SURVIVOR stamps on a note
-    /// file (rows NOT being undone that remain applied) so their on-disk stamps can be replayed after
-    /// the file is reverted to pristine. `path ↔ meeting` is 1:1 (each note has a unique `.md`), so a
-    /// meeting id keys the file's survivors.
-    pub fn supersessions_touching_meeting(
-        &self,
-        meeting_id: &str,
-    ) -> Result<Vec<crate::storage::models::SupersessionRow>> {
-        let conn = self.lock();
-        let mut stmt = conn
-            .prepare(
-                "SELECT id, superseding_meeting_id, source_meeting_id, entity, predicate,
-                        old_value, new_value, created_at, applied_at, source_pre_image,
-                        superseding_pre_image
-                   FROM supersessions
-                  WHERE source_meeting_id = ?1 OR superseding_meeting_id = ?1
-                  ORDER BY created_at ASC, id ASC",
-            )
-            .map_err(map_err)?;
-        let rows = stmt
-            .query_map(rusqlite::params![meeting_id], row_to_supersession)
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
-
-    /// RAW read: one supersession row by id (for apply/undo).
-    pub fn get_supersession(
-        &self,
-        id: &str,
-    ) -> Result<Option<crate::storage::models::SupersessionRow>> {
-        let conn = self.lock();
-        conn.query_row(
-            "SELECT id, superseding_meeting_id, source_meeting_id, entity, predicate,
-                    old_value, new_value, created_at, applied_at, source_pre_image,
-                    superseding_pre_image
-               FROM supersessions WHERE id = ?1",
-            rusqlite::params![id],
-            row_to_supersession,
-        )
-        .optional()
-        .map_err(map_err)
-    }
-
-    /// DURABLE-BEFORE-WRITE undo capture: persist the exact PRISTINE (pre-batch) pre-image bytes of
-    /// the note(s) a supersession is about to stamp, BEFORE the `.md` is written, so a crash between
-    /// the write and `mark_supersession_applied` still leaves a recoverable un-stamped pre-image. The
-    /// apply path resolves the bytes from a per-note-file pristine cache and calls this at most once
-    /// per field — and never re-snapshots a possibly-stamped file. `COALESCE` makes a `None` argument
-    /// a NO-OP for that column (it never clobbers an already-stored pristine backup — e.g. a retry
-    /// where the superseding note has since sealed). `applied_at` is left untouched (still NULL until
-    /// the write completes).
-    pub fn store_supersession_pre_images(
-        &self,
-        id: &str,
-        source_pre_image: Option<&[u8]>,
-        superseding_pre_image: Option<&[u8]>,
-    ) -> Result<()> {
-        let conn = self.lock();
-        conn.execute(
-            "UPDATE supersessions
-                SET source_pre_image = COALESCE(?2, source_pre_image),
-                    superseding_pre_image = COALESCE(?3, superseding_pre_image)
-              WHERE id = ?1",
-            rusqlite::params![id, source_pre_image, superseding_pre_image],
-        )
-        .map_err(map_err)?;
-        Ok(())
-    }
-
-    /// Mark a supersession APPLIED: stamp `applied_at` ONLY. The pre-images are captured separately
-    /// (and durably) by `store_supersession_pre_images` BEFORE the file write — so `applied_at` is the
-    /// LAST write, flipped only once the note(s) are safely stamped.
-    pub fn mark_supersession_applied(&self, id: &str, applied_at: &str) -> Result<()> {
-        let conn = self.lock();
-        conn.execute(
-            "UPDATE supersessions SET applied_at = ?2 WHERE id = ?1",
-            rusqlite::params![id, applied_at],
-        )
-        .map_err(map_err)?;
-        Ok(())
-    }
-
-    /// Clear the APPLIED state (undo): `applied_at` back to NULL and both pre-images dropped, once the
-    /// caller has restored the note bytes on disk. The pre-images are transient undo scratch — they do
-    /// not linger once the stamp is reverted.
-    pub fn clear_supersession_applied(&self, id: &str) -> Result<()> {
-        let conn = self.lock();
-        conn.execute(
-            "UPDATE supersessions
-                SET applied_at = NULL, source_pre_image = NULL, superseding_pre_image = NULL
-              WHERE id = ?1",
-            rusqlite::params![id],
-        )
-        .map_err(map_err)?;
-        Ok(())
-    }
 
     /// GATED `/people` personal-CRM rollup: one [`PersonCard`] per VISIBLE Person entity, built
     /// ENTIRELY from the existing gated graph/facts/commitment readers — NO new or ungated query.
@@ -10414,246 +8998,6 @@ impl Db {
         })
     }
 
-    // ── CROSS-MEETING USER MEMORY (Phase 3) ────────────────────────────────────
-    //
-    // User facts reuse the bitemporal `crate::facts::{Fact, FactOp}` shape and the PURE deterministic
-    // `reconcile_facts` core, but persist to the SEPARATE `user_facts` table (no entity FK). In the
-    // in-memory `Fact`/`NewFact` the `entity_id` field carries the USER-SCOPE SENTINEL
-    // (`crate::user_memory::USER_SCOPE`) so `reconcile_facts` keys on `(sentinel, subject, predicate)`
-    // — it is the reconcile key only, NOT a stored column. LOCK MODEL: `user_facts_all` is the
-    // INTERNAL un-gated reconcile input (pipeline-only, before any seal can hide rows, like
-    // `facts_for_entities`); every USER-FACING read goes through `list_user_facts_visible`, and a
-    // sealed meeting's user facts are PURGED on seal (`purge_user_facts_tx`).
-
-    /// ALL user facts (open + closed) — the reconcile input. INTERNAL: the un-gated lifecycle read
-    /// (the pipeline reconciles before any seal can hide rows), NOT a user-facing surface. Rows are
-    /// hydrated into `crate::facts::Fact` with `entity_id` set to the user-scope sentinel so the pure
-    /// `reconcile_facts` keys them correctly. Newest-recorded last (stable reconcile order).
-    pub fn user_facts_all(&self) -> Result<Vec<crate::facts::Fact>> {
-        let conn = self.lock();
-        let mut stmt = conn
-            .prepare(
-                "SELECT id, subject, predicate, object, valid_from, valid_to, recorded_at, \
-                        meeting_id, confidence \
-                   FROM user_facts ORDER BY recorded_at ASC, id ASC",
-            )
-            .map_err(map_err)?;
-        let rows = stmt.query_map([], row_to_user_fact).map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
-
-    /// Apply a batch of reconcile [`crate::facts::FactOp`]s to `user_facts` in ONE atomic transaction:
-    /// INSERT each `Add`, set `valid_to` on each `Invalidate` (only if still open — idempotent), skip
-    /// `NoOp`. A fresh UUID is minted per Add. The `entity_id` on an Add op is the user-scope sentinel
-    /// and is NOT persisted (there is no entity column). The whole batch commits or rolls back
-    /// together, so a crash mid-apply never leaves a half-reconciled store.
-    pub fn apply_user_fact_ops(&self, ops: &[crate::facts::FactOp]) -> Result<()> {
-        self.apply_user_fact_ops_inner(ops, None)
-    }
-
-    /// MEM-1: apply the reconcile ops AND, in the SAME atomic tx, record every `Invalidate`d
-    /// pre-existing fact under `import_meeting_id` (into `user_fact_import_supersedes`) so deleting the
-    /// synthetic Memory-Import meeting can REOPEN them. Used ONLY by `import_memories` — a normal
-    /// meeting's fact extraction anchors its Adds to the meeting itself and needs no reversible link
-    /// (deleting that meeting purges its Adds; it does not supersede OTHER meetings' facts). Recording
-    /// the closure link atomically with the closure means a crash can never leave a closed-but-unlinked
-    /// (permanently-lost-on-undo) fact.
-    pub fn apply_user_fact_ops_recording_import_supersedes(
-        &self,
-        ops: &[crate::facts::FactOp],
-        import_meeting_id: &str,
-    ) -> Result<()> {
-        self.apply_user_fact_ops_inner(ops, Some(import_meeting_id))
-    }
-
-    fn apply_user_fact_ops_inner(
-        &self,
-        ops: &[crate::facts::FactOp],
-        import_meeting_id: Option<&str>,
-    ) -> Result<()> {
-        use crate::facts::FactOp;
-        if ops.is_empty() {
-            return Ok(());
-        }
-        let mut conn = self.lock();
-        let tx = conn.transaction().map_err(map_err)?;
-        for op in ops {
-            match op {
-                FactOp::Add(nf) => {
-                    let id = uuid::Uuid::new_v4().to_string();
-                    tx.execute(
-                        "INSERT INTO user_facts \
-                           (id, subject, predicate, object, valid_from, valid_to, \
-                            recorded_at, meeting_id, confidence) \
-                         VALUES (?1, ?2, ?3, ?4, ?5, NULL, ?6, ?7, ?8)",
-                        rusqlite::params![
-                            id,
-                            nf.subject,
-                            nf.predicate,
-                            nf.object,
-                            nf.valid_from,
-                            nf.recorded_at,
-                            nf.meeting_id,
-                            nf.confidence,
-                        ],
-                    )
-                    .map_err(map_err)?;
-                }
-                FactOp::Invalidate { id, valid_to } => {
-                    let closed = tx
-                        .execute(
-                            "UPDATE user_facts SET valid_to = ?2 WHERE id = ?1 AND valid_to IS NULL",
-                            rusqlite::params![id, valid_to],
-                        )
-                        .map_err(map_err)?;
-                    // MEM-1: only record the reversible link when WE actually closed the row (the
-                    // `AND valid_to IS NULL` guard) AND this is an import run — so the undo reopens
-                    // exactly the facts THIS import superseded, nothing else.
-                    if closed > 0 {
-                        if let Some(imid) = import_meeting_id {
-                            tx.execute(
-                                "INSERT OR IGNORE INTO user_fact_import_supersedes
-                                   (import_meeting_id, superseded_fact_id, superseded_valid_to)
-                                 VALUES (?1, ?2, ?3)",
-                                rusqlite::params![imid, id, valid_to],
-                            )
-                            .map_err(map_err)?;
-                        }
-                    }
-                }
-                FactOp::NoOp => {}
-            }
-        }
-        tx.commit().map_err(map_err)?;
-        Ok(())
-    }
-
-    /// MEM-1 reversible-supersede UNDO (in an EXISTING tx, called from `delete_meeting`): when the
-    /// deleted meeting is a synthetic Memory-Import, REOPEN every pre-existing fact that import closed
-    /// (set `valid_to` back to NULL) — but ONLY where the row is still closed at the EXACT `valid_to`
-    /// we stamped (so a later legitimate re-supersession by a DIFFERENT meeting is never clobbered) —
-    /// then drop the link rows. Then the caller's `purge_user_facts_tx` deletes the import's OWN Adds.
-    /// A non-import meeting has no link rows ⇒ a no-op. This makes "delete the import ⇒ full undo"
-    /// restore the pre-existing memories instead of leaving them permanently closed.
-    fn reopen_import_superseded_facts_tx(
-        tx: &rusqlite::Transaction<'_>,
-        import_meeting_id: &str,
-    ) -> Result<()> {
-        // Reopen each still-matching superseded fact (guarded on the recorded valid_to).
-        tx.execute(
-            "UPDATE user_facts
-                SET valid_to = NULL
-              WHERE id IN (SELECT superseded_fact_id FROM user_fact_import_supersedes
-                            WHERE import_meeting_id = ?1)
-                AND valid_to = (SELECT superseded_valid_to FROM user_fact_import_supersedes s
-                                 WHERE s.import_meeting_id = ?1 AND s.superseded_fact_id = user_facts.id)",
-            rusqlite::params![import_meeting_id],
-        )
-        .map_err(map_err)?;
-        // Drop the link rows (the import is being deleted).
-        tx.execute(
-            "DELETE FROM user_fact_import_supersedes WHERE import_meeting_id = ?1",
-            rusqlite::params![import_meeting_id],
-        )
-        .map_err(map_err)?;
-        Ok(())
-    }
-
-    /// GATED read: the CURRENTLY-VALID (open) user facts whose SOURCE meeting is VISIBLE, newest
-    /// valid_from first. Visibility uses the SAME predicate as every other graph/MCP read
-    /// (`visibility_clause`): a user fact is visible iff its source meeting has a visible note (or no
-    /// note yet). A row with a NULL `meeting_id` is NOT visible — the INNER JOIN to `meetings` drops
-    /// it (fail-closed). This is the single user-facing user-fact read: it feeds BOTH the audit view
-    /// (`get_user_memory`) AND the injected memory brief, so a sealed-and-not-session-unlocked
-    /// meeting's user facts surface NOTHING and are injected into NO prompt. Only OPEN facts
-    /// (`valid_to IS NULL`) are returned — a forgotten/superseded fact is closed and excluded.
-    pub fn list_user_facts_visible(
-        &self,
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<crate::facts::Fact>> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        let sql = format!(
-            "SELECT uf.id, uf.subject, uf.predicate, uf.object, uf.valid_from, \
-                    uf.valid_to, uf.recorded_at, uf.meeting_id, uf.confidence \
-               FROM user_facts uf \
-               JOIN meetings m ON m.id = uf.meeting_id \
-              WHERE uf.valid_to IS NULL \
-                AND ( \
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id) \
-                   OR EXISTS ( \
-                        SELECT 1 FROM notes n \
-                         LEFT JOIN folders f ON f.id = n.folder_id \
-                         WHERE n.meeting_id = m.id AND {visible} \
-                      ) \
-                    ) \
-              ORDER BY uf.valid_from DESC, uf.id DESC"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt.query_map([], row_to_user_fact).map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
-
-    /// Brain v2 L2.2 — GATED, RELEVANCE-FILTERED read: the top-`k` CURRENTLY-VALID (open) user
-    /// facts matching `query` (BM25 over `fts_user_facts`, best first), restricted by EXACTLY the
-    /// same visibility predicate as [`Self::list_user_facts_visible`] (source-meeting
-    /// `visibility_clause`; NULL `meeting_id` fail-closed via the INNER JOIN). The query is defused
-    /// through [`fts_match_query_any`] (an OR of quoted literal CONTENT terms — stopwords and
-    /// <3-char tokens are dropped, so a natural-language question must match a fact on a real
-    /// content word, never on "the"/"is"), so raw user text can never raise an FTS syntax error; an
-    /// empty / punctuation-only / all-stopword query returns NO hits (the caller falls back to the
-    /// full list). Only OPEN facts (`valid_to IS NULL`) are returned.
-    pub fn search_user_facts_visible(
-        &self,
-        query: &str,
-        k: usize,
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<crate::facts::Fact>> {
-        let Some(match_expr) = fts_match_query_any(query) else {
-            return Ok(Vec::new());
-        };
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        let sql = format!(
-            "SELECT uf.id, uf.subject, uf.predicate, uf.object, uf.valid_from, \
-                    uf.valid_to, uf.recorded_at, uf.meeting_id, uf.confidence \
-               FROM fts_user_facts \
-               JOIN user_facts uf ON uf.rowid = fts_user_facts.rowid \
-               JOIN meetings m ON m.id = uf.meeting_id \
-              WHERE fts_user_facts MATCH ?1 \
-                AND uf.valid_to IS NULL \
-                AND ( \
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id) \
-                   OR EXISTS ( \
-                        SELECT 1 FROM notes n \
-                         LEFT JOIN folders f ON f.id = n.folder_id \
-                         WHERE n.meeting_id = m.id AND {visible} \
-                      ) \
-                    ) \
-              ORDER BY bm25(fts_user_facts) ASC, uf.valid_from DESC, uf.id DESC \
-              LIMIT ?2"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt
-            .query_map(
-                rusqlite::params![match_expr, k as i64],
-                row_to_user_fact,
-            )
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
 
     // ── Brain v2 L2.1 — memory consolidation store (see `crate::memory`) ─────────────────────────
     //
@@ -10770,36 +9114,6 @@ impl Db {
         )
         .map_err(map_err)?;
         Ok(path.flatten())
-    }
-
-    /// The persisted `facts.importance` assessments as `(fact_id, importance)` — the reflection
-    /// job's "already assessed" set for ENTITY facts (only never-assessed facts hit the reasoner,
-    /// so steady-state passes are LLM-free). Content-free read (ids + floats).
-    pub fn fact_importance_map(&self) -> Result<std::collections::HashMap<String, f64>> {
-        let conn = self.lock();
-        let mut stmt = conn
-            .prepare("SELECT id, importance FROM facts WHERE importance IS NOT NULL")
-            .map_err(map_err)?;
-        let rows = stmt
-            .query_map([], |r| Ok((r.get::<_, String>(0)?, r.get::<_, f64>(1)?)))
-            .map_err(map_err)?;
-        let mut out = std::collections::HashMap::new();
-        for r in rows {
-            let (id, imp) = r.map_err(map_err)?;
-            out.insert(id, imp);
-        }
-        Ok(out)
-    }
-
-    /// Persist the batch-assessed importance (1–10) of ONE entity fact (`facts.importance`).
-    pub fn set_fact_importance(&self, fact_id: &str, importance: f64) -> Result<()> {
-        let conn = self.lock();
-        conn.execute(
-            "UPDATE facts SET importance = ?2 WHERE id = ?1",
-            rusqlite::params![fact_id, importance],
-        )
-        .map_err(map_err)?;
-        Ok(())
     }
 
     /// Stamp the vault path a rollup was exported to (after the atomic `.md` write succeeded).
@@ -11029,117 +9343,6 @@ impl Db {
         Ok(n)
     }
 
-    /// FORGET one user fact by id (bitemporal INVALIDATE, never a silent delete): close the row at
-    /// `at` if it is still open. Idempotent (a already-closed row is untouched). History is preserved
-    /// — the fact simply stops being current, so it drops out of `list_user_facts_visible` and the
-    /// regenerated brief. Returns `true` iff a row was closed by this call.
-    pub fn forget_user_fact(&self, id: &str, at: &str) -> Result<bool> {
-        let conn = self.lock();
-        let n = conn
-            .execute(
-                "UPDATE user_facts SET valid_to = ?2 WHERE id = ?1 AND valid_to IS NULL",
-                rusqlite::params![id, at],
-            )
-            .map_err(map_err)?;
-        Ok(n > 0)
-    }
-
-    /// CLEAR all user memory: bitemporal-close EVERY currently-open user fact at `at` (invalidate,
-    /// never delete — closed history stays for the record). After this the brief regenerates empty and
-    /// the audit view is empty. Returns the number of facts closed.
-    pub fn clear_user_facts(&self, at: &str) -> Result<usize> {
-        let conn = self.lock();
-        let n = conn
-            .execute(
-                "UPDATE user_facts SET valid_to = ?1 WHERE valid_to IS NULL",
-                rusqlite::params![at],
-            )
-            .map_err(map_err)?;
-        Ok(n)
-    }
-
-    // The `audit_findings` / `audit_runs` CRUD (`insert_audit_finding_if_new` /
-    // `list_audit_finding_rows` / `get_audit_finding` / `resolve_audit_finding_row` /
-    // `delete_pending_audit_findings_for_run` / `count_pending_audit_findings` / `insert_audit_run` /
-    // `insert_scheduled_audit_run_claim` / `last_scheduled_audit_run_finished_at` /
-    // `delete_pending_audit_finding` / `purge_pending_audit_findings_tx` /
-    // `purge_all_pending_audit_findings_tx` / `purge_all_pending_audit_findings`) + `row_to_audit_finding`
-    // moved to `storage::audit_store` (God-file split) — still callable as inherent `db.method()`
-    // (and `Self::purge_*_pending_audit_findings_tx(&tx, ..)` from the seal paths) cross-file.
-
-    /// GATED: every OPEN fact whose source meeting is VISIBLE — the SAME meeting-visibility
-    /// predicate as [`Db::list_facts_visible`] (a NULL `meeting_id` is fail-closed via the INNER
-    /// JOIN). The audit's stale/contradiction substrate.
-    pub fn list_open_facts_visible(
-        &self,
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<crate::facts::Fact>> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        let sql = format!(
-            "SELECT ft.id, ft.entity_id, ft.subject, ft.predicate, ft.object, ft.valid_from, \
-                    ft.valid_to, ft.recorded_at, ft.meeting_id, ft.confidence \
-               FROM facts ft \
-               JOIN meetings m ON m.id = ft.meeting_id \
-              WHERE ft.valid_to IS NULL \
-                AND ( \
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id) \
-                   OR EXISTS ( \
-                        SELECT 1 FROM notes n \
-                         LEFT JOIN folders f ON f.id = n.folder_id \
-                         WHERE n.meeting_id = m.id AND {visible} \
-                      ) \
-                    ) \
-              ORDER BY ft.valid_from ASC, ft.id ASC"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt.query_map([], row_to_fact).map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
-
-    /// GATED: ONE meeting's FULL fact rows (open + closed), visible-predicate-gated exactly like
-    /// [`Db::list_facts_visible`] — a sealed-and-not-session-unlocked meeting returns EMPTY
-    /// (its facts are also purged on seal; this gate is defense-in-depth). Distinct from the
-    /// rendered-strings reader [`Db::facts_for_meeting_visible`]: the audit's staleness math
-    /// needs the `valid_to` axis, not a display string.
-    pub fn fact_rows_for_meeting_visible(
-        &self,
-        meeting_id: &str,
-        unlocked: &HashSet<String>,
-    ) -> Result<Vec<crate::facts::Fact>> {
-        let conn = self.lock();
-        let visible = visibility_clause("n", unlocked);
-        let sql = format!(
-            "SELECT ft.id, ft.entity_id, ft.subject, ft.predicate, ft.object, ft.valid_from, \
-                    ft.valid_to, ft.recorded_at, ft.meeting_id, ft.confidence \
-               FROM facts ft \
-               JOIN meetings m ON m.id = ft.meeting_id \
-              WHERE ft.meeting_id = ?1 \
-                AND ( \
-                      NOT EXISTS (SELECT 1 FROM notes nn WHERE nn.meeting_id = m.id) \
-                   OR EXISTS ( \
-                        SELECT 1 FROM notes n \
-                         LEFT JOIN folders f ON f.id = n.folder_id \
-                         WHERE n.meeting_id = m.id AND {visible} \
-                      ) \
-                    ) \
-              ORDER BY ft.valid_from ASC, ft.id ASC"
-        );
-        let mut stmt = conn.prepare(&sql).map_err(map_err)?;
-        let rows = stmt
-            .query_map(rusqlite::params![meeting_id], row_to_fact)
-            .map_err(map_err)?;
-        let mut out = Vec::new();
-        for r in rows {
-            out.push(r.map_err(map_err)?);
-        }
-        Ok(out)
-    }
-
     /// The DISTINCT entity ids mentioned by ONE meeting (the orphan pass's Jaccard substrate).
     /// Non-content metadata (opaque ids); the caller only ever passes ids already gated into its
     /// visible corpus. Crate-internal on purpose — not a general read surface.
@@ -11159,29 +9362,6 @@ impl Db {
             out.push(r.map_err(map_err)?);
         }
         Ok(out)
-    }
-
-    /// Is the SAME conflict pair already staged for Re-Truth review? True when a PENDING
-    /// (`applied_at IS NULL`) supersession row carries the same normalized predicate and the same
-    /// two meetings (either orientation) — the audit's contradiction pass then skips the pair
-    /// (one review surface per conflict).
-    pub fn pending_supersession_for_pair(
-        &self,
-        norm_predicate: &str,
-        meeting_a: &str,
-        meeting_b: &str,
-    ) -> Result<bool> {
-        let conn = self.lock();
-        conn.query_row(
-            "SELECT EXISTS(SELECT 1 FROM supersessions
-               WHERE applied_at IS NULL
-                 AND lower(trim(predicate)) = ?1
-                 AND ((source_meeting_id = ?2 AND superseding_meeting_id = ?3)
-                   OR (source_meeting_id = ?3 AND superseding_meeting_id = ?2)))",
-            rusqlite::params![norm_predicate, meeting_a, meeting_b],
-            |r| r.get(0),
-        )
-        .map_err(map_err)
     }
 
     /// Whether ONE authored note is VISIBLE (open or session-unlocked folder) — the lightweight
@@ -11347,27 +9527,14 @@ pub(crate) fn backlink_sort_key(b: &BacklinkSource) -> i64 {
 pub(crate) type LinkRowRaw =
     (i64, String, String, String, String, String, String, String, f64, i64);
 
-/// One VISIBLE note/document node row for the full-brain graph, pre-typed: `(kind, id, label, date)`
-/// (`date` = the `updated_at`/`created_at` epoch-ms as a string). Aliased to keep
-/// [`Db::full_graph_content_nodes`]'s return under clippy's type-complexity bar.
-type FullGraphContentNode = (FullGraphNodeKind, String, String, Option<String>);
-
-/// One raw `links` row for the full-brain graph, PRE-gating: `(src_kind, src_id, dst_kind, dst_id,
-/// edge_type, score, status)` (ids/kinds/metadata only — no content). Aliased to keep
-/// [`Db::full_graph_links`]'s return under clippy's type-complexity bar.
-type FullGraphLinkRow = (String, String, String, String, String, f64, String);
-
-/// One entity→meeting MENTION edge for the full-brain graph: `(entity_id, meeting_id, weight)`.
-/// Aliased to keep [`Db::entity_meeting_mentions_visible`]'s `(Vec<_>, truncated)` return under
-/// clippy's type-complexity bar (PR-9 F2 added the `truncated` flag).
-type FullGraphMentionEdge = (String, String, i64);
-
 /// Per-kind render cap for the full-brain graph (`build_full_graph`, DESIGN §PR-4). Meetings, notes,
 /// and documents are each capped independently (entities keep their own 500 cap inside
 /// `list_entities_visible`) so the payload stays bounded on a large vault; visibility is enforced in
 /// WHERE, so the cap trims magnitude only — it can never widen what is visible. Mirrors the
-/// `MAX_GRAPH_EDGES`/`MAX_VISIBLE_ENTITIES` posture, disclosed via `total_visible_nodes`.
-const MAX_FULL_GRAPH_PER_KIND: usize = 500;
+/// `MAX_GRAPH_EDGES`/`MAX_VISIBLE_ENTITIES` posture, disclosed via `total_visible_nodes`. Consumed by
+/// `storage::graph_store::build_full_graph`; kept here (with the graph tests in `mod tests`) so both
+/// see it via one `pub(crate)` symbol.
+pub(crate) const MAX_FULL_GRAPH_PER_KIND: usize = 500;
 
 /// Bound for the `links`-derived edge leg of the full-brain graph (`full_graph_links`, PR-9 F2).
 /// `links` is the fastest-growing edge table (every wikilink/companion/semantic-suggestion is a row)
@@ -11375,40 +9542,7 @@ const MAX_FULL_GRAPH_PER_KIND: usize = 500;
 /// exist for was unenforced on its highest-cardinality leg. Strongest-score edges survive the cut;
 /// the trim is DISCLOSED via `FullGraphData::edges_truncated`. Comfortably above what a laid-out
 /// ≤140-node scene can render, so it trims magnitude only on a very large vault.
-const MAX_FULL_GRAPH_LINK_EDGES: usize = 4000;
-
-/// Convert an epoch-MILLISECONDS timestamp to an RFC3339 UTC string (PR-9 F4: unify the
-/// full-graph node `date` format — meetings already carry ISO `started_at`; notes/docs store
-/// epoch-ms). `None` for an out-of-range value (never a panic — a bad date hint is dropped, not
-/// fatal). Uses the same `chrono` dependency already in use across storage.
-fn epoch_ms_to_rfc3339(ms: i64) -> Option<String> {
-    chrono::DateTime::<chrono::Utc>::from_timestamp_millis(ms).map(|dt| dt.to_rfc3339())
-}
-
-/// Map a persisted `links.src_kind`/`dst_kind` string onto the STABLE full-graph node kind
-/// (`meeting`/`note`/`document`). `None` for anything unknown (a corrupt row → its edge is dropped).
-/// PR-9 F4: returns the typed [`FullGraphNodeKind`] (not a bare `&str`) so the caller can carry the
-/// endpoint kinds onto `FullGraphEdge` — the FE then matches endpoints by `(kind, id)`, not bare id,
-/// closing a cross-kind id-collision mismatch. Its `.as_str()` remains the visible-node-set key.
-fn full_graph_link_node_kind(kind: &str) -> Option<FullGraphNodeKind> {
-    match kind {
-        "meeting" => Some(FullGraphNodeKind::Meeting),
-        "note" => Some(FullGraphNodeKind::Note),
-        "document" => Some(FullGraphNodeKind::Document),
-        _ => None,
-    }
-}
-
-/// Map a persisted `links.edge_type` onto the full-graph edge kind. `None` for an unknown type (the
-/// edge is dropped). co_occurrence/mention are computed edges, never stored, so they are not here.
-fn full_graph_edge_kind_from_type(edge_type: &str) -> Option<FullGraphEdgeKind> {
-    match edge_type {
-        "wikilink" => Some(FullGraphEdgeKind::Wikilink),
-        "companion" => Some(FullGraphEdgeKind::Companion),
-        "semantic" => Some(FullGraphEdgeKind::Semantic),
-        _ => None,
-    }
-}
+pub(crate) const MAX_FULL_GRAPH_LINK_EDGES: usize = 4000;
 
 pub(crate) fn visibility_clause(_alias: &str, unlocked: &HashSet<String>) -> String {
     let mut clause = String::from("(f.locked IS NULL OR f.locked = 0");
@@ -11650,12 +9784,12 @@ fn date_clause(date_filter: Option<&(String, String)>) -> String {
 /// Minimum entity-name length (in chars) eligible for QUERY→ENTITY resolution (GraphRAG-lite).
 /// Names shorter than this are too noisy as whole-query tokens (e.g. 2-letter initials) and are
 /// never resolved.
-const MIN_ENTITY_NAME_LEN: usize = 3;
+pub(crate) const MIN_ENTITY_NAME_LEN: usize = 3;
 
 /// Lowercase + tokenize on Unicode non-alphanumeric boundaries (Polish-safe via
 /// `char::is_alphanumeric`). Empty tokens are dropped. Used by the deterministic QUERY→ENTITY
 /// resolver so matching is on whole tokens, never arbitrary substrings.
-fn tokenize_lower(s: &str) -> Vec<String> {
+pub(crate) fn tokenize_lower(s: &str) -> Vec<String> {
     s.to_lowercase()
         .split(|c: char| !c.is_alphanumeric())
         .filter(|t| !t.is_empty())
@@ -11667,7 +9801,7 @@ fn tokenize_lower(s: &str) -> Vec<String> {
 /// `query_tokens` (already lowercased). So "atlas" matches the query "atlas status" but NOT the
 /// substring "atlasian"; a multi-token name like "anna kowalska" matches only when both tokens are
 /// adjacent and ordered. Guards the entity expansion against spurious substring noise.
-fn name_matches_query_tokens(query_tokens: &[String], name_ci: &str) -> bool {
+pub(crate) fn name_matches_query_tokens(query_tokens: &[String], name_ci: &str) -> bool {
     let name_tokens: Vec<&str> = name_ci
         .split(|c: char| !c.is_alphanumeric())
         .filter(|t| !t.is_empty())
@@ -11684,7 +9818,7 @@ fn name_matches_query_tokens(query_tokens: &[String], name_ci: &str) -> bool {
 // Return `Result<Meeting>` (not raw rusqlite::Result) so the status-parse error
 // is surfaced as an AppError; `query_row(...).optional()?.transpose()` flattens it.
 
-fn row_to_meeting(row: &Row<'_>) -> rusqlite::Result<Result<Meeting>> {
+pub(crate) fn row_to_meeting(row: &Row<'_>) -> rusqlite::Result<Result<Meeting>> {
     // Read every column as a rusqlite result first (so `?` here yields rusqlite::Error),
     // then fold the status-string parse (which yields AppError) into the inner Result.
     let id: String = row.get(0)?;
@@ -11713,40 +9847,7 @@ fn row_to_meeting(row: &Row<'_>) -> rusqlite::Result<Result<Meeting>> {
 // `row_to_audit_finding` moved to `storage::audit_store` (God-file split) alongside the
 // `audit_findings` readers that are its only callers.
 
-/// Map a `facts` row (column order matches every facts SELECT) to a [`crate::facts::Fact`].
-fn row_to_fact(row: &Row<'_>) -> rusqlite::Result<crate::facts::Fact> {
-    Ok(crate::facts::Fact {
-        id: row.get(0)?,
-        entity_id: row.get(1)?,
-        subject: row.get(2)?,
-        predicate: row.get(3)?,
-        object: row.get(4)?,
-        valid_from: row.get(5)?,
-        valid_to: row.get(6)?,
-        recorded_at: row.get(7)?,
-        meeting_id: row.get(8)?,
-        confidence: row.get(9)?,
-    })
-}
 
-/// Map a `user_facts` row (column order: id, subject, predicate, object, valid_from, valid_to,
-/// recorded_at, meeting_id, confidence — NO entity column) to a [`crate::facts::Fact`], stamping the
-/// user-scope sentinel into `entity_id` so the pure `reconcile_facts` keys the row correctly. The
-/// sentinel is a reconcile key only, never persisted.
-fn row_to_user_fact(row: &Row<'_>) -> rusqlite::Result<crate::facts::Fact> {
-    Ok(crate::facts::Fact {
-        id: row.get(0)?,
-        entity_id: crate::user_memory::USER_SCOPE.to_string(),
-        subject: row.get(1)?,
-        predicate: row.get(2)?,
-        object: row.get(3)?,
-        valid_from: row.get(4)?,
-        valid_to: row.get(5)?,
-        recorded_at: row.get(6)?,
-        meeting_id: row.get(7)?,
-        confidence: row.get(8)?,
-    })
-}
 
 // `row_to_brief_schedule` / `row_to_brief_run` moved to `storage::brief_store` (God-file split)
 // alongside the `brief_schedules` / `brief_runs` readers that are their only callers.
@@ -11754,24 +9855,6 @@ fn row_to_user_fact(row: &Row<'_>) -> rusqlite::Result<crate::facts::Fact> {
 // `row_to_mcp_server` moved to `storage::mcp_store` (God-file split) alongside the `mcp_servers`
 // readers that are its only callers.
 
-/// Map a `supersessions` row (id, superseding_meeting_id, source_meeting_id, entity, predicate,
-/// old_value, new_value, created_at, applied_at, source_pre_image, superseding_pre_image) to a row
-/// struct.
-fn row_to_supersession(row: &Row<'_>) -> rusqlite::Result<crate::storage::models::SupersessionRow> {
-    Ok(crate::storage::models::SupersessionRow {
-        id: row.get(0)?,
-        superseding_meeting_id: row.get(1)?,
-        source_meeting_id: row.get(2)?,
-        entity: row.get(3)?,
-        predicate: row.get(4)?,
-        old_value: row.get(5)?,
-        new_value: row.get(6)?,
-        created_at: row.get(7)?,
-        applied_at: row.get(8)?,
-        source_pre_image: row.get(9)?,
-        superseding_pre_image: row.get(10)?,
-    })
-}
 
 /// Escape LIKE wildcards so user input is matched literally (paired with `ESCAPE '\'`).
 fn escape_like(s: &str) -> String {
@@ -11829,7 +9912,7 @@ fn fts_match_query(q: &str) -> Option<String> {
 /// unicode61 has no stopword list and BM25 cannot rescue a hit set whose only members are stopword
 /// matches). Empty / punctuation-only / all-stopword input yields `None` — the caller's fallback
 /// owns that case.
-fn fts_match_query_any(q: &str) -> Option<String> {
+pub(crate) fn fts_match_query_any(q: &str) -> Option<String> {
     let mut terms: Vec<String> = Vec::new();
     let mut cur = String::new();
     for ch in q.chars() {
@@ -22405,7 +20488,10 @@ mod graph_tests {
     //! and the Sink-B `locked`-gate (disk-truth, not session-unlock) for the vault stub mirror.
 
     use super::*;
-    use crate::storage::models::{EntityKind, Folder, Meeting, MeetingStatus, NoteRecord};
+    use crate::storage::models::{
+        EntityKind, Folder, FullGraphEdgeKind, FullGraphNodeKind, FullGraphOpts, GraphEdge, Meeting,
+        MeetingStatus, NoteRecord,
+    };
 
     const TEST_DEK: &str = "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef";
