+++ src-tauri/src/storage/mod.rs
@@ -1,8 +1,11 @@
+pub(crate) mod ask_store;
 pub(crate) mod audit_store;
 pub(crate) mod brief_store;
 pub mod db;
 pub(crate) mod egress_store;
+pub(crate) mod facts_store;
 pub(crate) mod folders_store;
+pub(crate) mod graph_store;
 pub(crate) mod links;
 pub(crate) mod mcp_store;
 pub mod migration;
